Move SVG example to ImageScreen and move return statement
