+++ demo/src/screens/incubatorScreens/index.js
@@ -5,5 +5,4 @@ export function registerScreens(registrar) {
     gestureHandlerRootHOC(require('./TouchableOpacityScreen').default));
   registrar('unicorn.components.IncubatorTextFieldScreen', () => require('./IncubatorTextFieldScreen').default);
   registrar('unicorn.incubator.WheelPickerScreen', () => gestureHandlerRootHOC(require('./WheelPickerScreen').default));
-  registrar('unicorn.incubator.SvgImageScreen', () => require('./SvgImageScreen').default);
 }
